zsbl: plat: sg2042: detect board type dynamically
The conf.ini contains board configuration information. The convention is
that the conf.ini must start at byte 0 with a section called sophgo-config,
without any comments. In addition, the current conf.ini contains the
[devicetree] section, which has two keys, namely name and addr.

The conf.ini is saved in SPI Flash1 and will be added to EEPROM in the
future.

Signed-off-by: jingyu.li01 <jingyu.li01@sophgo.com>